+++ src/app/app.component.ts
@@ -10,7 +10,7 @@ import { IconComponent } from './shared/components/icon/icon.component';
 import { FieldLibrarySidebarComponent } from './features/field-library-sidebar/field-library-sidebar.component';
 import { PreferencesModalComponent } from './features/modals/preferences/preferences.component';
 import { PresetDefinitionsModalComponent } from './features/modals/preset-definitions/preset-definitions.component';
-import { PreviewPanelComponent } from './features/preview-panel/preview-panel.component';
+import { CeePreviewComponent } from './features/cee-preview/cee-preview.component';
 import { FieldTypePickerComponent } from './features/field-type-picker/field-type-picker.component';
 import { FieldDesignerComponent } from './features/field-designer/field-designer.component';
 import { FieldCardComponent } from './features/field-card/field-card.component';
@@ -27,7 +27,7 @@ import { CedarExportAccordionsComponent } from './features/cedar-export-accordio
     FieldLibrarySidebarComponent,
     PreferencesModalComponent,
     PresetDefinitionsModalComponent,
-    PreviewPanelComponent,
+    CeePreviewComponent,
     FieldTypePickerComponent,
     FieldDesignerComponent,
     FieldCardComponent,